+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Hyperedge.normalized().
 - New package graphbrain.meaning, for meaning extractors.
 - New package graphbrain.readers, for text readers from various sources.
+- New package graphbrain.corefs, for coreferences resolution.
 
 ### Changed
 - Sequences keep track of their size, no need to specify position when adding edge to sequence.